\name{ex3n300a.data}
\docType{data}
\alias{ex3n300a.data}
\title{Simulated QTL dataset in QTLCartographer format}
\description{Simulated data for 300 backcross progeny similar to the
  example used for testing QTL co-location (Ball (2007).
}
\usage{ex3n200a.data}
\references{
Ball, R. D. 2007: Quantifying evidence for candidate gene
polymorphisms---Bayesian analysis combining sequence-specific and QTL
co-location information. Genetics 177: 2399--2416. 
}
\keyword{datasets}