\name{print.summary.bicreg.qtl}
\alias{print.summary.bicreg.qtl}
\title{Print summary of bicreg.qtl objects}
\description{
  \code{print} method for \code{summary.bicreg.qtl}
  objects}
\usage{
\method{print}{summary.bicreg.qtl}(x,...)
}
\arguments{
  \item{x}{object of class 'bicreg.summ'}
  \item{...}{further arguments to genetic print methods}
}
\author{Rod Ball (\email{rod.ball@scionresearch.com})}
\keyword{print}