\name{summary.bicreg.qtl}
\alias{summary.bicreg.qtl}
\title{Summary of bicreg.qtl objects}
\description{
  \code{summary} method for \code{bicreg.qtl} objects}
\usage{
\method{summary}{bicreg.qtl}(object,...)
}
\arguments{
  \item{object}{object of class 'bicreg.qtl'}
  \item{...}{further arguments to other summary methods}
}
\author{Rod Ball (\email{rod.ball@AT@scionresearch.com})}
\keyword{print}